Rename: 8팀 공통 컴포넌트 폴더 경로 수정
